Fix some small bugs

+++ src/main/java/vn/elca/training/web/ApplicationController.java
@@ -75,10 +75,10 @@ public class ApplicationController {
     public String multilingual(Locale locale, Model model) {
         return "search";
     }
-    
-    @RequestMapping(value = "/login", method = RequestMethod.GET)
+
+    @RequestMapping(value = "/loginUrl", method = RequestMethod.GET)
     public String loginPage() {
-        return "login";
+        return "loginpage";
     }
 
     /**
@@ -137,12 +137,12 @@ public class ApplicationController {
     @RequestMapping("/paging/{page}")
     @ResponseBody
     List<Project> paging(HttpSession session, @PathVariable Integer page, Model model, Locale locale) {
-        String searchCriteria = (session.getAttribute("TEXT_SEARCH_CRITERIA") != null
-                ? session.getAttribute("TEXT_SEARCH_CRITERIA").toString() : "all");
-        String status = (session.getAttribute("STATUS_SEARCH_CRITERIA") != null
-                ? session.getAttribute("STATUS_SEARCH_CRITERIA").toString() : "-1");
-        String sortOrdering = (session.getAttribute("SORT_ORDERING") != null
-                ? session.getAttribute("SORT_ORDERING").toString() : "asc");
+        String searchCriteria = (session.getAttribute("TEXT_SEARCH_CRITERIA") != null ? session.getAttribute(
+                "TEXT_SEARCH_CRITERIA").toString() : "all");
+        String status = (session.getAttribute("STATUS_SEARCH_CRITERIA") != null ? session.getAttribute(
+                "STATUS_SEARCH_CRITERIA").toString() : "-1");
+        String sortOrdering = (session.getAttribute("SORT_ORDERING") != null ? session.getAttribute("SORT_ORDERING")
+                .toString() : "asc");
         String sortName = (session.getAttribute("SORT_NAME") != null ? session.getAttribute("SORT_NAME").toString()
                 : "id");
         // return the search result
@@ -164,10 +164,10 @@ public class ApplicationController {
     @ResponseBody
     List<Project> sort(HttpSession session, @PathVariable String colName, Model model, Locale locale) {
         // process session variables
-        String searchCriteria = (session.getAttribute("TEXT_SEARCH_CRITERIA") != null
-                ? session.getAttribute("TEXT_SEARCH_CRITERIA").toString() : "all");
-        String status = (session.getAttribute("STATUS_SEARCH_CRITERIA") != null
-                ? session.getAttribute("STATUS_SEARCH_CRITERIA").toString() : "-1");
+        String searchCriteria = (session.getAttribute("TEXT_SEARCH_CRITERIA") != null ? session.getAttribute(
+                "TEXT_SEARCH_CRITERIA").toString() : "all");
+        String status = (session.getAttribute("STATUS_SEARCH_CRITERIA") != null ? session.getAttribute(
+                "STATUS_SEARCH_CRITERIA").toString() : "-1");
         String sortOrdering = "asc";
         if (session.getAttribute("SORT_ORDERING") != null) {
             sortOrdering = session.getAttribute("SORT_ORDERING").toString();
@@ -194,8 +194,14 @@ public class ApplicationController {
     @RequestMapping("/delete")
     @ResponseBody
     String delete(@RequestParam(value = "prjIds[]") List<Long> prjIds) {
+        Project verifyPrj;
         for (Long id : prjIds) {
-            this.projectService.delete(id);
+            verifyPrj = this.projectService.getById(String.valueOf(id));
+            if ("NEW".equals(verifyPrj.getStatus().toUpperCase())) {
+                this.projectService.delete(id);
+            } else {
+                return "fail";
+            }
         }
         // callback to main() to re-display [home] view
         return "success";
@@ -281,8 +287,8 @@ public class ApplicationController {
     }
 
     /**
-     * Convert from [String] type to [Date] type for [finishingDate] parameter.
-     * The input formatting of [finishingDate] should be [dd/MM/yyyy"].
+     * Convert from [String] type to [Date] type for [finishingDate] parameter. The input formatting of [finishingDate]
+     * should be [dd/MM/yyyy"].
      * 
      * @param binder
      */
